Wire up load button to load the file
It looks like this was probably just an oversight! Was going to file an
issue for it but figured if the cause was what I assumed it would be a
fairly easy fix.

Standard caveat that this contribution is being done in my personal time
and does not imply support for prechoster from anti software software
club.